Rename test-clients.
