+++ Cargo.toml
@@ -22,15 +22,17 @@ members = [
 	"core/cli",
 	"core/client",
 	"core/client/db",
-	"core/consensus/common",
 	"core/consensus/aura",
 	"core/consensus/babe",
+	"core/consensus/common",
 	"core/consensus/rhd",
 	"core/consensus/slots",
 	"core/executor",
 	"core/finality-grandpa",
 	"core/finality-grandpa/primitives",
+	"core/inherents",
 	"core/keyring",
+	"core/keystore",
 	"core/network",
 	"core/panic-handler",
 	"core/primitives",
@@ -46,13 +48,13 @@ members = [
 	"core/sr-std",
 	"core/sr-version",
 	"core/state-machine",
-	"core/test-runtime",
 	"core/telemetry",
-	"core/trie",
-	"core/keystore",
+	"core/test-client",
+	"core/test-runtime",
+	"core/test-runtime/client",
 	"core/transaction-pool",
 	"core/transaction-pool/graph",
-	"core/inherents",
+	"core/trie",
 	"core/util/fork-tree",
 	"srml/support",
 	"srml/support/procedural",
