+++ core/basic-authorship/Cargo.toml
@@ -17,4 +17,4 @@ transaction_pool = { package = "substrate-transaction-pool", path = "../../core/
 substrate-telemetry = { path = "../telemetry" }
 
 [dev-dependencies]
-test-client = { package = "substrate-test-client", path = "../../core/test-client" }
+test-client = { package = "substrate-test-runtime-client", path = "../../core/test-runtime/client" }
