+++ core/client/db/Cargo.toml
@@ -25,7 +25,7 @@ consensus_common = { package = "substrate-consensus-common", path = "../../conse
 
 [dev-dependencies]
 substrate-keyring = { path = "../../keyring" }
-test-client = { package = "substrate-test-client", path = "../../test-client" }
+test-client = { package = "substrate-test-runtime-client", path = "../../test-runtime/client" }
 env_logger = { version = "0.6" }
 
 [features]
