+++ core/network/Cargo.toml
@@ -30,16 +30,16 @@ peerset = { package = "substrate-peerset", path = "../../core/peerset" }
 tokio-timer = "0.2.11"
 tokio = { version = "0.1.11", optional = true }
 keyring = { package = "substrate-keyring", path = "../../core/keyring", optional = true }
-test_client = { package = "substrate-test-client", path = "../../core/test-client", optional = true }
+test-client = { package = "substrate-test-runtime-client", path = "../../core/test-runtime/client", optional = true }
 void = "1.0"
 
 [dev-dependencies]
 env_logger = { version = "0.6" }
 keyring = { package = "substrate-keyring", path = "../../core/keyring" }
-test_client = { package = "substrate-test-client", path = "../../core/test-client" }
+test-client = { package = "substrate-test-runtime-client", path = "../../core/test-runtime/client" }
 consensus = { package = "substrate-consensus-common", path = "../../core/consensus/common", features = ["test-helpers"] }
 tokio = "0.1.11"
 
 [features]
 default = []
-test-helpers = ["keyring", "test_client", "consensus/test-helpers", "tokio"]
+test-helpers = ["keyring", "test-client", "consensus/test-helpers", "tokio"]
