+++ core/test-client/Cargo.toml
@@ -5,20 +5,15 @@ authors = ["Parity Technologies <admin@parity.io>"]
 edition = "2018"
 
 [dependencies]
-generic-test-client = { package = "substrate-generic-test-client", path = "./generic" }
+client = { package = "substrate-client", path = "../client" }
+client-db = { package = "substrate-client-db", path = "../client/db", features = ["test-helpers"] }
+consensus = { package = "substrate-consensus-common", path = "../consensus/common" }
+executor = { package = "substrate-executor", path = "../executor" }
+futures = { version = "0.1.27" }
+hash-db = "0.12"
+keyring = { package = "substrate-keyring", path = "../keyring" }
+parity-codec = "3.5.1"
 primitives = { package = "substrate-primitives", path = "../primitives" }
-runtime = { package = "substrate-test-runtime", path = "../test-runtime", default-features = false }
 runtime_primitives = { package = "sr-primitives", path = "../sr-primitives" }
+state_machine = { package = "substrate-state-machine", path = "../state-machine" }
 
-[features]
-default = [
-	"include-wasm-blob",
-	"std",
-]
-std = [
-	"runtime/std",
-]
-# If enabled, the WASM blob is added to the `GenesisConfig`.
-include-wasm-blob = [
-	"runtime/include-wasm-blob",
-]
