--- core/test-client/generic/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "substrate-generic-test-client"
-version = "2.0.0"
-authors = ["Parity Technologies <admin@parity.io>"]
-edition = "2018"
-
-[dependencies]
-client = { package = "substrate-client", path = "../../client" }
-client-db = { package = "substrate-client-db", path = "../../client/db", features = ["test-helpers"] }
-consensus = { package = "substrate-consensus-common", path = "../../consensus/common" }
-executor = { package = "substrate-executor", path = "../../executor" }
-futures = { version = "0.1.17" }
-hash-db = "0.12"
-keyring = { package = "substrate-keyring", path = "../../../core/keyring" }
-parity-codec = "3.5.1"
-primitives = { package = "substrate-primitives", path = "../../primitives" }
-runtime_primitives = { package = "sr-primitives", path = "../../sr-primitives" }
-state_machine = { package = "substrate-state-machine", path = "../../state-machine" }
-
