+++ core/test-client/src/client_ext.rs
