+++ core/test-client/src/lib.rs
@@ -1,4 +1,4 @@
-// Copyright 2018-2019 Parity Technologies (UK) Ltd.
+// Copyright 2019 Parity Technologies (UK) Ltd.
 // This file is part of Substrate.
 
 // Substrate is free software: you can redistribute it and/or modify
@@ -18,198 +18,253 @@
 
 #![warn(missing_docs)]
 
-#[cfg(feature = "include-wasm-blob")]
-pub mod trait_tests;
+pub mod client_ext;
 
-mod block_builder_ext;
+pub use client::{ExecutionStrategies, blockchain, backend, self};
+pub use client_db::{Backend, self};
+pub use client_ext::ClientExt;
+pub use consensus;
+pub use executor::{NativeExecutor, self};
+pub use keyring::{sr25519::Keyring as AuthorityKeyring, AccountKeyring};
+pub use primitives::Blake2Hasher;
+pub use runtime_primitives::{StorageOverlay, ChildrenStorageOverlay};
+pub use state_machine::ExecutionStrategy;
 
-pub use block_builder_ext::BlockBuilderExt;
-pub use generic_test_client::*;
-pub use runtime;
-
-use runtime::genesismap::{GenesisConfig, additional_storage_with_genesis};
-use runtime_primitives::traits::{Block as BlockT, Header as HeaderT, Hash as HashT};
-
-/// A prelude to import in tests.
-pub mod prelude {
-	// Trait extensions
-	pub use super::{BlockBuilderExt, DefaultTestClientBuilderExt, TestClientBuilderExt, ClientExt};
-	// Client structs
-	pub use super::{
-		TestClient, TestClientBuilder, Backend, LightBackend,
-		Executor, LightExecutor, LocalExecutor, NativeExecutor,
-	};
-	// Keyring
-	pub use super::{AccountKeyring, AuthorityKeyring};
-}
-
-#[cfg(feature = "include-wasm-blob")]
-mod local_executor {
-	#![allow(missing_docs)]
-	use runtime;
-	use crate::executor::native_executor_instance;
-	// FIXME #1576 change the macro and pass in the `BlakeHasher` that dispatch needs from here instead
-	native_executor_instance!(
-		pub LocalExecutor,
-		runtime::api::dispatch,
-		runtime::native_version,
-		include_bytes!("../../test-runtime/wasm/target/wasm32-unknown-unknown/release/substrate_test_runtime.compact.wasm")
-	);
-}
-
-/// Native executor used for tests.
-#[cfg(feature = "include-wasm-blob")]
-pub use local_executor::LocalExecutor;
-
-/// Test client database backend.
-pub type Backend = generic_test_client::Backend<runtime::Block>;
-
-/// Test client executor.
-#[cfg(feature = "include-wasm-blob")]
-pub type Executor = client::LocalCallExecutor<
-	Backend,
-	NativeExecutor<LocalExecutor>,
->;
+use std::sync::Arc;
+use std::collections::HashMap;
+use futures::future::FutureResult;
+use hash_db::Hasher;
+use primitives::storage::well_known_keys;
+use runtime_primitives::traits::{
+	Block as BlockT, NumberFor
+};
+use client::LocalCallExecutor;
 
 /// Test client light database backend.
-pub type LightBackend = generic_test_client::LightBackend<runtime::Block>;
-
-/// Test client light executor.
-#[cfg(feature = "include-wasm-blob")]
-pub type LightExecutor = client::light::call_executor::RemoteOrLocalCallExecutor<
-	runtime::Block,
-	LightBackend,
-	client::light::call_executor::RemoteCallExecutor<
-		client::light::blockchain::Blockchain<
-			client_db::light::LightStorage<runtime::Block>,
-			LightFetcher
-		>,
-		LightFetcher
-	>,
-	client::LocalCallExecutor<
-		client::light::backend::Backend<
-			client_db::light::LightStorage<runtime::Block>,
-			LightFetcher,
-			Blake2Hasher
-		>,
-		NativeExecutor<LocalExecutor>
-	>
+pub type LightBackend<Block> = client::light::backend::Backend<
+	client_db::light::LightStorage<Block>,
+	LightFetcher,
+	Blake2Hasher,
 >;
 
-/// Parameters of test-client builder with test-runtime.
-#[derive(Default)]
-pub struct GenesisParameters {
-	support_changes_trie: bool,
+/// Test client light fetcher.
+pub struct LightFetcher;
+
+/// A genesis storage initialisation trait.
+pub trait GenesisInit: Default {
+	/// Construct genesis storage.
+	fn genesis_storage(&self) -> (StorageOverlay, ChildrenStorageOverlay);
 }
 
-impl generic_test_client::GenesisInit for GenesisParameters {
+impl GenesisInit for () {
 	fn genesis_storage(&self) -> (StorageOverlay, ChildrenStorageOverlay) {
-		let mut storage = genesis_config(self.support_changes_trie).genesis_map();
-
-		let state_root = <<<runtime::Block as BlockT>::Header as HeaderT>::Hashing as HashT>::trie_root(
-			storage.clone().into_iter()
-		);
-		let block: runtime::Block = client::genesis::construct_genesis_block(state_root);
-		storage.extend(additional_storage_with_genesis(&block));
-
-		(storage, Default::default())
+		Default::default()
 	}
 }
 
-/// A `TestClient` with `test-runtime` builder.
-pub type TestClientBuilder<E, B> = generic_test_client::TestClientBuilder<E, B, GenesisParameters>;
-
-/// Test client type with `LocalExecutor` and generic Backend.
-pub type Client<B> = client::Client<
-	B,
-	client::LocalCallExecutor<B, executor::NativeExecutor<LocalExecutor>>,
-	runtime::Block,
-	runtime::RuntimeApi,
->;
-
-/// A test client with default backend.
-pub type TestClient = Client<Backend>;
-
-/// A `TestClientBuilder` with default backend and executor.
-pub trait DefaultTestClientBuilderExt: Sized {
-	/// Create new `TestClientBuilder`
-	fn new() -> Self;
+/// A builder for creating a test client instance.
+pub struct TestClientBuilder<Executor, Backend, G: GenesisInit = ()> {
+	execution_strategies: ExecutionStrategies,
+	genesis_init: G,
+	child_storage_extension: HashMap<Vec<u8>, Vec<(Vec<u8>, Vec<u8>)>>,
+	backend: Arc<Backend>,
+	_executor: std::marker::PhantomData<Executor>,
 }
 
-impl DefaultTestClientBuilderExt for TestClientBuilder<
+impl<Block, Executor> Default for TestClientBuilder<
 	Executor,
-	Backend,
-> {
-	fn new() -> Self {
+	Backend<Block>,
+> where
+	Block: BlockT<Hash=<Blake2Hasher as Hasher>::Out>,
+{
+	fn default() -> Self {
 		Self::with_default_backend()
 	}
 }
 
-/// A `test-runtime` extensions to `TestClientBuilder`.
-pub trait TestClientBuilderExt<B>: Sized {
-	/// Enable or disable support for changes trie in genesis.
-	fn set_support_changes_trie(self, support_changes_trie: bool) -> Self;
-
-	/// Build the test client.
-	fn build(self) -> Client<B> {
-		self.build_with_longest_chain().0
+impl<Block, Executor, G: GenesisInit> TestClientBuilder<
+	Executor,
+	Backend<Block>,
+	G,
+> where
+	Block: BlockT<Hash=<Blake2Hasher as Hasher>::Out>,
+{
+	/// Create new `TestClientBuilder` with default backend.
+	pub fn with_default_backend() -> Self {
+		let backend = Arc::new(Backend::new_test(std::u32::MAX, std::u64::MAX));
+		Self::with_backend(backend)
 	}
-
-	/// Build the test client and longest chain selector.
-	fn build_with_longest_chain(self) -> (Client<B>, client::LongestChain<B, runtime::Block>);
 }
 
-impl<B> TestClientBuilderExt<B> for TestClientBuilder<
-	client::LocalCallExecutor<B, executor::NativeExecutor<LocalExecutor>>,
-	B
-> where
-	B: client::backend::Backend<runtime::Block, Blake2Hasher>,
-{
-	fn set_support_changes_trie(mut self, support_changes_trie: bool) -> Self {
-		self.genesis_init_mut().support_changes_trie = support_changes_trie;
+impl<Executor, Backend, G: GenesisInit> TestClientBuilder<
+	Executor,
+	Backend,
+	G,
+> {
+	/// Create a new instance of the test client builder.
+	pub fn with_backend(backend: Arc<Backend>) -> Self {
+		TestClientBuilder {
+			backend,
+			execution_strategies: ExecutionStrategies::default(),
+			child_storage_extension: Default::default(),
+			genesis_init: Default::default(),
+			_executor: Default::default(),
+		}
+	}
+
+	/// Alter the genesis storage parameters.
+	pub fn genesis_init_mut(&mut self) -> &mut G {
+		&mut self.genesis_init
+	}
+
+	/// Extend child storage
+	pub fn add_child_storage(
+		mut self,
+		key: impl AsRef<[u8]>,
+		child_key: impl AsRef<[u8]>,
+		value: impl AsRef<[u8]>,
+	) -> Self {
+		let entry = self.child_storage_extension.entry(key.as_ref().to_vec()).or_insert_with(Vec::new);
+		entry.push((child_key.as_ref().to_vec(), value.as_ref().to_vec()));
 		self
 	}
 
-	fn build_with_longest_chain(self) -> (Client<B>, client::LongestChain<B, runtime::Block>) {
-		self.build_with_native_executor(None)
+	/// Set the execution strategy that should be used by all contexts.
+	pub fn set_execution_strategy(
+		mut self,
+		execution_strategy: ExecutionStrategy
+	) -> Self {
+		self.execution_strategies = ExecutionStrategies {
+			syncing: execution_strategy,
+			importing: execution_strategy,
+			block_construction: execution_strategy,
+			offchain_worker: execution_strategy,
+			other: execution_strategy,
+		};
+		self
+	}
+
+	/// Build the test client with the given native executor.
+	pub fn build_with_executor<Block, RuntimeApi>(
+		self,
+		executor: Executor,
+	) -> (
+		client::Client<
+			Backend,
+			Executor,
+			Block,
+			RuntimeApi,
+		>,
+		client::LongestChain<Backend, Block>,
+	) where
+		Executor: client::CallExecutor<Block, Blake2Hasher>,
+		Backend: client::backend::Backend<Block, Blake2Hasher>,
+		Block: BlockT<Hash=<Blake2Hasher as Hasher>::Out>,
+	{
+
+		let storage = {
+			let mut storage = self.genesis_init.genesis_storage();
+
+			// Add some child storage keys.
+			for (key, value) in self.child_storage_extension {
+				storage.1.insert(
+					well_known_keys::CHILD_STORAGE_KEY_PREFIX.iter().cloned().chain(key).collect(),
+					value.into_iter().collect(),
+				);
+			}
+
+			storage
+		};
+
+		let client = client::Client::new(
+			self.backend.clone(),
+			executor,
+			storage,
+			self.execution_strategies
+		).expect("Creates new client");
+
+		let longest_chain = client::LongestChain::new(self.backend);
+
+		(client, longest_chain)
 	}
 }
 
-fn genesis_config(support_changes_trie: bool) -> GenesisConfig {
-	GenesisConfig::new(support_changes_trie, vec![
-		AuthorityKeyring::Alice.into(),
-		AuthorityKeyring::Bob.into(),
-		AuthorityKeyring::Charlie.into(),
-	], vec![
-		AccountKeyring::Alice.into(),
-		AccountKeyring::Bob.into(),
-		AccountKeyring::Charlie.into(),
-	],
-		1000
-	)
+impl<E, Backend, G: GenesisInit> TestClientBuilder<
+	client::LocalCallExecutor<Backend, executor::NativeExecutor<E>>,
+	Backend,
+	G,
+> {
+	/// Build the test client with the given native executor.
+	pub fn build_with_native_executor<Block, RuntimeApi, I>(
+		self,
+		executor: I,
+	) -> (
+		client::Client<
+			Backend,
+			client::LocalCallExecutor<Backend, executor::NativeExecutor<E>>,
+			Block,
+			RuntimeApi
+		>,
+		client::LongestChain<Backend, Block>,
+	) where
+		I: Into<Option<executor::NativeExecutor<E>>>,
+		E: executor::NativeExecutionDispatch,
+		Backend: client::backend::Backend<Block, Blake2Hasher>,
+		Block: BlockT<Hash=<Blake2Hasher as Hasher>::Out>,
+	{
+		let executor = executor.into().unwrap_or_else(|| executor::NativeExecutor::new(None));
+		let executor = LocalCallExecutor::new(self.backend.clone(), executor);
+
+		self.build_with_executor(executor)
+	}
 }
 
-/// Creates new client instance used for tests.
-#[cfg(feature = "include-wasm-blob")]
-pub fn new() -> Client<Backend> {
-	TestClientBuilder::new().build()
-}
+impl<Block: BlockT> client::light::fetcher::Fetcher<Block> for LightFetcher {
+	type RemoteHeaderResult = FutureResult<Block::Header, client::error::Error>;
+	type RemoteReadResult = FutureResult<Option<Vec<u8>>, client::error::Error>;
+	type RemoteCallResult = FutureResult<Vec<u8>, client::error::Error>;
+	type RemoteChangesResult = FutureResult<Vec<(NumberFor<Block>, u32)>, client::error::Error>;
+	type RemoteBodyResult = FutureResult<Vec<Block::Extrinsic>, client::error::Error>;
 
-/// Creates new light client instance used for tests.
-#[cfg(feature = "include-wasm-blob")]
-pub fn new_light() -> client::Client<LightBackend, LightExecutor, runtime::Block, runtime::RuntimeApi> {
-	use std::sync::Arc;
+	fn remote_header(
+		&self,
+		_request: client::light::fetcher::RemoteHeaderRequest<Block::Header>,
+	) -> Self::RemoteHeaderResult {
+		unimplemented!("not (yet) used in tests")
+	}
 
-	let storage = client_db::light::LightStorage::new_test();
-	let blockchain = Arc::new(client::light::blockchain::Blockchain::new(storage));
-	let backend = Arc::new(LightBackend::new(blockchain.clone()));
-	let executor = NativeExecutor::new(None);
-	let fetcher = Arc::new(LightFetcher);
-	let remote_call_executor = client::light::call_executor::RemoteCallExecutor::new(blockchain.clone(), fetcher);
-	let local_call_executor = client::LocalCallExecutor::new(backend.clone(), executor);
-	let call_executor = LightExecutor::new(backend.clone(), remote_call_executor, local_call_executor);
+	fn remote_read(
+		&self,
+		_request: client::light::fetcher::RemoteReadRequest<Block::Header>,
+	) -> Self::RemoteReadResult {
+		unimplemented!("not (yet) used in tests")
+	}
 
-	TestClientBuilder::with_backend(backend)
-		.build_with_executor(call_executor)
-		.0
+	fn remote_read_child(
+		&self,
+		_request: client::light::fetcher::RemoteReadChildRequest<Block::Header>,
+	) -> Self::RemoteReadResult {
+		unimplemented!("not (yet) used in tests")
+	}
+
+	fn remote_call(
+		&self,
+		_request: client::light::fetcher::RemoteCallRequest<Block::Header>,
+	) -> Self::RemoteCallResult {
+		unimplemented!("not (yet) used in tests")
+	}
+
+	fn remote_changes(
+		&self,
+		_request: client::light::fetcher::RemoteChangesRequest<Block::Header>,
+	) -> Self::RemoteChangesResult {
+		unimplemented!("not (yet) used in tests")
+	}
+
+	fn remote_body(
+		&self,
+		_request: client::light::fetcher::RemoteBodyRequest<Block::Header>,
+	) -> Self::RemoteBodyResult {
+		unimplemented!("not (yet) used in tests")
+	}
 }
