+++ node/executor/Cargo.toml
@@ -17,7 +17,7 @@ node-primitives = { path = "../primitives" }
 node-runtime = { path = "../runtime" }
 
 [dev-dependencies]
-test-client = { package = "substrate-generic-test-client", path = "../../core/test-client/generic/" }
+test-client = { package = "substrate-test-client", path = "../../core/test-client" }
 keyring = { package = "substrate-keyring", path = "../../core/keyring" }
 runtime_primitives = { package = "sr-primitives", path = "../../core/sr-primitives" }
 runtime_support = { package = "srml-support", path = "../../srml/support" }
